backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
python-dotenv==1.0.0
langchain==0.1.4
langchain-anthropic==0.1.1
langchain-core==0.1.17
anthropic==0.18.1
pydantic==2.5.3
pydantic-settings==2.1.0
PyMuPDF==1.23.8
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
aiofiles==23.2.1
